Move create_apk_elf_path() helper into new apk module
Move the create_apk_elf_path() helper function into the newly created
apk module, to make it accessible from outside the symbolization logic.
