+++ src/symbolize/symbolizer.rs
@@ -13,6 +13,8 @@ use std::path::Path;
 use std::path::PathBuf;
 use std::rc::Rc;
 
+#[cfg(feature = "apk")]
+use crate::apk::create_apk_elf_path;
 #[cfg(feature = "breakpad")]
 use crate::breakpad::BreakpadResolver;
 use crate::elf::ElfParser;
@@ -111,29 +113,6 @@ impl Debug for DebugMapsEntry<'_> {
             .field(stringify!(path), &path.display())
             .finish()
     }
-}
-
-
-#[cfg(feature = "apk")]
-fn create_apk_elf_path(apk: &Path, elf: &Path) -> Result<PathBuf> {
-    let mut extension = apk
-        .extension()
-        .unwrap_or_else(|| OsStr::new("apk"))
-        .to_os_string();
-    // Append '!' to indicate separation from archive internal contents
-    // that follow. This is an Android convention.
-    let () = extension.push("!");
-
-    let mut apk = apk.to_path_buf();
-    if !apk.set_extension(extension) {
-        return Err(Error::with_invalid_data(format!(
-            "path {} is not valid",
-            apk.display()
-        )))
-    }
-
-    let path = apk.join(elf);
-    Ok(path)
 }
 
 
@@ -1525,18 +1504,6 @@ mod tests {
         let () = entries.for_each(|entry| {
             assert_ne!(format!("{:?}", DebugMapsEntry(&entry.unwrap())), "");
         });
-    }
-
-    /// Check that we can create a path to an ELF inside an APK as expected.
-    #[test]
-    fn elf_apk_path_creation() {
-        let apk = Path::new("/root/test.apk");
-        let elf = Path::new("subdir/libc.so");
-        let path = create_apk_elf_path(apk, elf).unwrap();
-        assert_eq!(path, Path::new("/root/test.apk!/subdir/libc.so"));
-
-        let err = create_apk_elf_path(Path::new(""), elf).unwrap_err();
-        assert_eq!(err.kind(), ErrorKind::InvalidData);
     }
 
     /// Check that we can correctly construct the source code path to a symbol.
